Move ConfigValueLocation to a separate file
